backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
ibm-watson
ibm-cloud-sdk-core
python-jose[cryptography]
bcrypt
